requirements.txt
appdirs==1.4.4
APScheduler==3.9.1.post1
atomicwrites==1.4.0
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
coloredlogs==15.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dicttoxml==1.7.4
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
fpdf==1.7.2
gunicorn==20.1.0
humanfriendly==10.0
idna==3.2
iniconfig==1.1.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
json2xml==3.6.0
MarkupSafe==2.0.1
marshmallow==3.12.2
marshmallow-sqlalchemy==0.26.1
mysqlclient==2.0.3
packaging==21.0
Pillow==8.3.1
pluggy==1.0.0
py==1.10.0
pyasn1==0.4.8
pycryptodome==3.10.1
PyJWT==2.1.0
pymysql==1.0.2
pyparsing==2.4.7
pyreadline3==3.3
pytest==6.2.5
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
qrcode==7.2
redis==4.0.2
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
rsa==4.7.2
sendgrid==6.9.7
six==1.16.0
SQLAlchemy==1.4.20
starkbank-ecdsa==2.2.0
toml==0.10.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.13.3
zeep==4.0.0
pip~=22.2.2
mysql-connector-python~=8.0.33
Flask-Limiter==3.8.0